Moved Config(uration) class
