+++ service/src/test/java/dk/dbc/search/solrdocstore/AgencyLibraryTypeBeanIT.java
@@ -3,7 +3,6 @@ package dk.dbc.search.solrdocstore;
 import dk.dbc.search.solrdocstore.openagency.libraryrules.LibraryRuleException;
 import dk.dbc.search.solrdocstore.openagency.libraryrules.LibraryRulesProxy;
 import dk.dbc.search.solrdocstore.openagency.libraryrules.OpenAgencyClient;
-import dk.dbc.search.solrdocstore.openagency.libraryrules.OpenAgencyConfig;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
@@ -25,7 +24,7 @@ public class AgencyLibraryTypeBeanIT extends JpaSolrDocStoreIntegrationTester {
         bean.entityManager = em;
         bean.proxy = new LibraryRulesProxy();
         OpenAgencyClient openAgencyClient = new OpenAgencyClient();
-        OpenAgencyConfig openAgencyConfig = new OpenAgencyConfig();
+        Config openAgencyConfig = new Config();
         openAgencyConfig.loadProperties();
         openAgencyClient.setConfig(openAgencyConfig);
         bean.proxy.setOaclient(openAgencyClient);
